requirements.txt
amqp==5.0.6
asgiref==3.4.1
autoflake==1.4
billiard==3.6.4.0
black==21.9b0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==3.9.2
invoke==1.6.0
isort==5.9.3
kombu==5.1.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
regex==2021.8.28
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
